refactor(api): move CampaignAuthorization and prescriber-role-repository to prescription/campaign
