+++ api/src/prescription/campaign/domain/read-models/CampaignAuthorization.js
